refactor: split RuboCop report shaping from execution
Keep the Ruby sensor below the repository's file-size gate by separating
process execution from report parsing and finding shaping.
